Added needed methods
