+++ src/main/java/frc/robot/Robot.java
@@ -73,7 +73,6 @@ public class Robot extends TimedRobot {
     visionRight.addFilteredPose();
     visionLeft.addFilteredPose();
 
-
     DogLog.log("CoralPosition/isCoralInFunnel", CoralPosition.isCoralInFunnel());
     DogLog.log("CoralPosition/isCoralInTootsieSlide", CoralPosition.isCoralInTootsieSlide());
   }
@@ -173,8 +172,7 @@ public class Robot extends TimedRobot {
 
   /** This function is called periodically during operator control. */
   @Override
-  public void teleopPeriodic() {
-  }
+  public void teleopPeriodic() {}
 
   @Override
   public void testInit() {
